fixing up wagmi
