+++ src/features/colinks/BuyOrSellCoLinks.tsx
@@ -5,7 +5,7 @@ import { getBalance } from '@wagmi/core';
 import { ethers } from 'ethers';
 import { defaultAvailableChains } from 'features/DecentSwap/config';
 import { DecentSwap } from 'features/DecentSwap/DecentSwap';
-import { wagmiChain, wagmiConfig } from 'features/wagmi/config';
+import { wagmiConfig } from 'features/wagmi/config';
 import { useQuery } from 'react-query';
 import { NavLink } from 'react-router-dom';
 import type { CSS } from 'stitches.config';
@@ -19,6 +19,7 @@ import { Check, Link2 } from '../../icons/__generated';
 import { client } from '../../lib/gql/client';
 import { Button, Flex, Panel, Text } from '../../ui';
 import { sendAndTrackTx } from '../../utils/viem/contractHelpers';
+import { wagmiChain } from '../wagmi/chains';
 import { BridgeButton } from 'components/BridgeButton';
 import { OptimismBridgeButton } from 'components/OptimismBridgeButton';
 import { OrBar } from 'components/OrBar';
