+++ src/features/web3/rpcs.ts
@@ -1,0 +1,9 @@
+import { VITE_FE_ALCHEMY_API_KEY } from '../../config/env';
+
+export const ETHEREUM_RPC_URL = `https://eth-mainnet.g.alchemy.com/v2/${VITE_FE_ALCHEMY_API_KEY}`;
+export const OPTIMISM_SEPOLIA_RPC_URL = `https://opt-sepolia.g.alchemy.com/v2/${VITE_FE_ALCHEMY_API_KEY}`;
+export const ETHEREUM_SEPOLIA_RPC_URL = `https://eth-sepolia.g.alchemy.com/v2/${VITE_FE_ALCHEMY_API_KEY}`;
+export const POLYGON_RPC_URL = `https://polygon-mainnet.g.alchemy.com/v2/${VITE_FE_ALCHEMY_API_KEY}`;
+export const ARBITRUM_RPC_URL = `https://arb-mainnet.g.alchemy.com/v2/${VITE_FE_ALCHEMY_API_KEY}`;
+export const BASE_RPC_URL = `https://base-mainnet.g.alchemy.com/v2/${VITE_FE_ALCHEMY_API_KEY}`; // TODO: Refactor these to be defined in one place across entire app
+export const OPTIMISM_RPC_URL = `https://opt-mainnet.g.alchemy.com/v2/${VITE_FE_ALCHEMY_API_KEY}`;
